requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
pydantic>=2.0.0
pyyaml>=6.0
streamlit>=1.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
networkx>=3.0
graphviz>=0.20
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0

# Enhanced document parsing support
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.0
plotly>=5.17.0
